refactor(main): decompose src/main/index.ts into focused modules
- extract setupProcessListeners, setupIpcHandlers, Pianola relearn
  construction, and Cadenza/plugin-host-view wiring into their own modules
- preserve exact call order, the dead-but-harmless raw mainWindow field,
  and module-eval-time IPC registration timing
