+++ src/main/cadenza-bridge/index.ts
@@ -1,0 +1,3 @@
+export { createCadenzaDelivery } from './delivery';
+export { registerCadenzaIpcHandlers } from './ipc';
+export type { CadenzaBridgeDependencies, CadenzaIpcDependencies } from './types';
