+++ src/main/index.ts
@@ -23,10 +23,8 @@ import { readFile } from 'fs/promises';
 import { ProcessManager } from './process-manager';
 import { WebServer } from './web-server';
 import { AgentDetector } from './agents';
-import { getAgentDefinition } from './agents/definitions';
-import { DEFAULT_CONTEXT_WINDOWS, FALLBACK_CONTEXT_WINDOW } from '../shared/agentConstants';
+import { createAgentConfigLookup } from './agents/agent-config-lookup';
 import { shouldDropSentryEvent } from '../shared/sentryFilters';
-import type { AgentId } from '../shared/agentIds';
 import {
 	initGlobalHotkey,
 	setGlobalShowHotkey,
@@ -36,11 +34,8 @@ import { CueEngine } from './cue/cue-engine';
 import { createCueSupervisorHooks } from './cue/cue-first-party';
 import { PianolaSupervisor } from './pianola/pianola-supervisor';
 import { PianolaRelearnScheduler } from './pianola/pianola-relearn-scheduler';
-import { runRelearnJob } from './pianola/pianola-relearn';
-import { readRules, writeSuggestions, getProfile } from './pianola/pianola-store-main';
-import type { DecisionPair } from '../shared/pianola/transcript-mining';
-import type { PianolaRule } from '../shared/pianola/types';
-import { spawn, execFile, type ChildProcess } from 'child_process';
+import { createPianolaLifecycle } from './pianola/pianola-lifecycle';
+import { execFile } from 'child_process';
 import { PluginManager } from './plugins/plugin-manager';
 import { resolveTrustedKeys } from '../shared/plugins/publisher-keys';
 import { seedBundledPlugins } from './plugins/bundled-plugins';
@@ -59,8 +54,8 @@ import {
 	type PluginSessionMetadata,
 	type PluginTabMetadata,
 } from './plugins/plugin-host-handlers';
-import { PluginHostViewRegistry, type HostViewMutation } from './plugins/plugin-host-view-registry';
-import { forwardPluginHostViewToRenderer } from './plugins/plugin-host-view-forwarder';
+import { createCadenzaDelivery, registerCadenzaIpcHandlers } from './cadenza-bridge';
+import { createPluginHostViewBridge } from './plugin-host-view-bridge';
 import { ActionGuard } from './plugins/action-guard';
 import { PluginKvStore } from './plugins/plugin-kv-store';
 import { PluginEventBusImpl } from './plugins/plugin-event-bus';
@@ -100,14 +95,9 @@ import {
 	type OpenedConsentWindow,
 } from './plugins/consent-window';
 import { configureCueTelemetry } from './cue/cue-telemetry';
-import {
-	executeCuePrompt,
-	recordCueHistoryEntry,
-	stopCueRun,
-	getCueProcessList,
-} from './cue/cue-executor';
+import { executeCuePrompt, recordCueHistoryEntry, stopCueRun } from './cue/cue-executor';
 import { executeCueShell, stopCueShellRun } from './cue/cue-shell-executor';
-import { executeCueCli, stopCueCliRun, resolveMaestroCliScriptPath } from './cue/cue-cli-executor';
+import { executeCueCli, stopCueCliRun } from './cue/cue-cli-executor';
 import { executeCueNotify } from './cue/cue-notify-executor';
 import { getAgentDisplayName } from '../shared/agentMetadata';
 import { logger } from './utils/logger';
@@ -126,120 +116,33 @@ import {
 	getWindowStateStore,
 	getClaudeSessionOriginsStore,
 	getAgentSessionOriginsStore,
-	getSshRemoteById,
 } from './stores';
 import { runSettingsMigrations } from './stores/migrations';
 import {
-	registerGitHandlers,
-	registerAutorunHandlers,
-	registerPlaybooksHandlers,
-	registerHistoryHandlers,
-	registerAgentsHandlers,
-	registerProcessHandlers,
-	registerPersistenceHandlers,
-	registerSystemHandlers,
-	registerClaudeHandlers,
-	registerAgentSessionsHandlers,
-	registerGroupChatHandlers,
-	registerDebugHandlers,
-	registerSpeckitHandlers,
-	registerOpenSpecHandlers,
-	registerBmadHandlers,
-	registerContextHandlers,
-	registerMarketplaceHandlers,
-	registerStatsHandlers,
-	registerCueStatsHandlers,
-	registerDocumentGraphHandlers,
-	registerSshRemoteHandlers,
-	registerFilesystemHandlers,
-	registerAttachmentsHandlers,
-	registerWebHandlers,
 	ensureCliServer,
 	startCliDiscoveryWatchdog,
 	stopCliDiscoveryWatchdog,
-	registerLeaderboardHandlers,
-	registerNotificationsHandlers,
-	registerSymphonyHandlers,
-	registerTabNamingHandlers,
-	registerAgentErrorHandlers,
-	registerDirectorNotesHandlers,
-	registerCrossAgentHandlers,
-	registerCueHandlers,
-	registerCueBackupHandlers,
-	registerWakatimeHandlers,
-	registerFeedbackHandlers,
-	registerMaestroCliHandlers,
-	registerPromptsHandlers,
-	registerMemoryHandlers,
-	registerPianolaHandlers,
-	registerPluginsHandlers,
-	registerAgentRunHandlers,
-	registerCoworkingHandlers,
-	registerBrowserSessionHandlers,
-	registerWindowsHandlers,
-	wireWindowRegistryBroadcast,
-	wireEmptySecondaryWindowAutoClose,
-	setupLoggerEventForwarding,
 	cleanupAllGroomingSessions,
 	getActiveGroomingSessionCount,
 } from './ipc/handlers';
-import { startCoworkingBridge, stopCoworkingBridge } from './coworking/coworking-bridge';
-import { ensureCoworkingServerScript } from './coworking/coworking-server-paths';
-import { resolveSessionFromPidWalk } from './coworking/pid-resolution';
-import { initializeStatsDB, closeStatsDB, getStatsDB, wireMultiWindowTelemetry } from './stats';
-import { groupChatEmitters } from './ipc/handlers/groupChat';
-import {
-	routeModeratorResponse,
-	routeAgentResponse,
-	setGetSessionsCallback,
-	setGetCustomEnvVarsCallback,
-	setGetAgentConfigCallback,
-	setGetModeratorSettingsCallback,
-	setSshStore,
-	setGetCustomShellPathCallback,
-	markParticipantResponded,
-	spawnModeratorSynthesis,
-	getGroupChatReadOnlyState,
-	respawnParticipantWithRecovery,
-	clearActiveParticipantTaskSession,
-	clearModeratorResponseTimeout,
-} from './group-chat/group-chat-router';
+import { setupIpcHandlers } from './ipc/bootstrap';
+import { stopCoworkingBridge } from './coworking/coworking-bridge';
+import { initializeStatsDB, closeStatsDB } from './stats';
 import { createSshRemoteStoreAdapter } from './utils/ssh-remote-resolver';
-import { updateParticipant, loadGroupChat, updateGroupChat } from './group-chat/group-chat-storage';
 import { stopSessionCleanup } from './group-chat/group-chat-moderator';
-import { needsSessionRecovery, initiateSessionRecovery } from './group-chat/session-recovery';
 import { initializePrompts, getPrompt, savePrompt } from './prompt-manager';
 import { captureException } from './utils/sentry';
-import { initializeSessionStorages } from './storage';
 import { resolveToFilePath, configureImageStore } from './storage/session-image-store';
 import { CONCERTO_HTML_SCHEME } from '../shared/concerto-html';
 import { createConcertoHtmlResponse } from './concerto-html';
 import { MEDIA_SCHEME } from '../shared/mediaTypes';
 import { handleMediaStreamRequest } from './media/media-stream';
-import { initializeOutputParsers } from './parsers';
-import { calculateContextTokens } from './parsers/usage-aggregator';
-import {
-	DEMO_MODE,
-	DEMO_DATA_PATH,
-	REGEX_MODERATOR_SESSION,
-	REGEX_MODERATOR_SESSION_TIMESTAMP,
-	REGEX_AI_SUFFIX,
-	REGEX_AI_TAB_ID,
-	REGEX_BATCH_SESSION,
-	REGEX_SYNOPSIS_SESSION,
-	debugLog,
-} from './constants';
+import { DEMO_MODE, DEMO_DATA_PATH } from './constants';
 // initAutoUpdater is now used by window-manager.ts (Phase 4 refactoring)
 import { checkWslEnvironment } from './utils/wslDetector';
 import { setupDeepLinkHandling, flushPendingDeepLink } from './deep-links';
 // Extracted modules (Phase 1 refactoring)
-import { parseParticipantSessionId } from './group-chat/session-parser';
-import { extractTextFromStreamJson } from './group-chat/output-parser';
-import {
-	appendToGroupChatBuffer,
-	getGroupChatBufferedOutput,
-	clearGroupChatBuffer,
-} from './group-chat/output-buffer';
+import { wireProcessListeners } from './process-listeners-wiring';
 // Phase 2 refactoring - dependency injection
 import { createSafeSend, isWebContentsAvailable } from './utils/safe-send';
 import { capabilitySnapshots, createSnapshotBroadcaster } from './agents/capability-snapshot';
@@ -251,29 +154,19 @@ import {
 	createSettingsWatcher,
 	createWindowManager,
 	createQuitHandler,
-	deliverCadenzaToHud,
-	deliverCadenzaToExistingHud,
 	closeCadenzaHudWindow,
-	getCadenzaHudWindow,
 	type QuitHandler,
 } from './app-lifecycle';
 // Multi-window registry (single source of truth for window<->session ownership)
 import { WindowRegistry } from './window-registry';
 // Multi-window startup restore: turn the persisted MultiWindowState back into
 // window-creation specs (pruning agents that no longer exist).
-import {
-	planWindowRestore,
-	pickFocusWindowSpec,
-	saveWindowState,
-} from './window-state-persistence';
+import { planWindowRestore, pickFocusWindowSpec } from './window-state-persistence';
 import type { WindowState as SharedWindowState } from '../shared/window-types';
-// Phase 3 refactoring - process listeners
-import { setupProcessListeners as setupProcessListenersModule } from './process-listeners';
 import { setupAgentRunCapture } from './agent-run/setup-capture-listener';
 import { setAgentRunSink } from './agent-run/broadcast';
 import { startAgentRunStoreWatcher } from './agent-run/store-watcher';
 import { setupAgentRunRecovery } from './agent-run/setup-recovery';
-import { setupWakaTimeListener } from './process-listeners/wakatime-listener';
 import { WakaTimeManager } from './wakatime-manager';
 import { MaestroCliManager } from './maestro-cli-manager';
 import {
@@ -497,14 +390,8 @@ const windowStateStore = getWindowStateStore();
 const claudeSessionOriginsStore = getClaudeSessionOriginsStore();
 const agentSessionOriginsStore = getAgentSessionOriginsStore();
 
-function getAgentConfigForAgent(agentId: string): Record<string, any> {
-	const allConfigs = agentConfigsStore.get('configs', {});
-	return allConfigs[agentId] || {};
-}
-
-function getCustomEnvVarsForAgent(agentId: string): Record<string, string> | undefined {
-	return getAgentConfigForAgent(agentId).customEnvVars as Record<string, string> | undefined;
-}
+const { getAgentConfigForAgent, getCustomEnvVarsForAgent } =
+	createAgentConfigLookup(agentConfigsStore);
 
 // Note: History storage is now handled by HistoryManager which uses per-session files
 // in the history/ directory. The legacy maestro-history.json file is migrated automatically.
@@ -530,74 +417,6 @@ let pluginEventBus: PluginEventBusImpl | null = null;
 let noteSessionActivatedInPersistence: ((sessionId: string) => void) | null = null;
 let usageRefreshScheduler: UsageRefreshScheduler | null = null;
 let interactiveReplayController: InteractiveReplayController<ProcessSpawnConfig> | null = null;
-
-/** Cap on decision pairs the scheduled re-learn pulls from the CLI per run. */
-const RELEARN_MAX_PAIRS = 100_000;
-
-/**
- * Mine the installed CLIs' native transcripts into a decision corpus by spawning
- * the existing `pianola learn --json` crawler (the single source of transcript
- * discovery + parsing) and parsing its `pairs`. Rejects on spawn/exit/parse
- * failure so a failed mine leaves the previously staged suggestions untouched.
- */
-function mineDecisionPairsViaCli(): Promise<DecisionPair[]> {
-	const cliScriptPath = resolveMaestroCliScriptPath();
-	return new Promise<DecisionPair[]>((resolve, reject) => {
-		let child: ChildProcess;
-		try {
-			child = spawn(
-				process.execPath,
-				[cliScriptPath, 'pianola', 'learn', '--json', '--max-pairs', String(RELEARN_MAX_PAIRS)],
-				{
-					env: {
-						...process.env,
-						// In packaged Electron, process.execPath is the app binary, not
-						// Node; without this it would launch the app instead of the CLI.
-						ELECTRON_RUN_AS_NODE: '1',
-						MAESTRO_CLI_JS: cliScriptPath,
-					},
-					stdio: ['ignore', 'pipe', 'pipe'],
-				}
-			);
-		} catch (err) {
-			reject(err instanceof Error ? err : new Error(String(err)));
-			return;
-		}
-		let stdout = '';
-		let stderr = '';
-		child.stdout?.setEncoding('utf8');
-		child.stdout?.on('data', (d: string) => {
-			stdout += d;
-		});
-		child.stderr?.setEncoding('utf8');
-		child.stderr?.on('data', (d: string) => {
-			stderr += d;
-		});
-		child.on('error', (err) => reject(err));
-		child.on('exit', (code) => {
-			if (code !== 0) {
-				reject(new Error(`pianola learn exited ${code ?? 'null'}: ${stderr.trim().slice(0, 200)}`));
-				return;
-			}
-			try {
-				const parsed = JSON.parse(stdout) as { pairs?: unknown };
-				resolve(Array.isArray(parsed.pairs) ? (parsed.pairs as DecisionPair[]) : []);
-			} catch (err) {
-				reject(err instanceof Error ? err : new Error(String(err)));
-			}
-		});
-	});
-}
-
-/**
- * Read the user's live rules and global decision-profile markdown for the
- * re-learn baseline. A missing or malformed profiles file degrades to an empty
- * baseline (getProfile already returns a well-formed empty result), so the job
- * stages a fresh draft rather than crashing.
- */
-function readExistingForRelearn(): { rules: PianolaRule[]; profile: string } {
-	return { rules: readRules(), profile: getProfile().entry?.profile ?? '' };
-}
 
 // Create safeSend with dependency injection (Phase 2 refactoring).
 // Broadcasts to EVERY open window, not just the primary one - see the
@@ -678,59 +497,21 @@ const cadenzaHudDeps = {
 	windowRegistry,
 };
 
-/**
- * Route a cadenza payload to the HUD window (creating it lazily). Returns
- * false when there's no main window to parent it, so the caller can fall back
- * to the in-app renderer.
- */
-function deliverCadenza(payload: Parameters<typeof deliverCadenzaToHud>[2]): boolean {
-	if (!mainWindow) return false;
-	// Concerto is an opt-in Encore feature: don't spawn the HUD window (or
-	// route anything) unless the user enabled it in Extensions.
-	if (store.get('encoreFeatures')?.concerto !== true) return false;
-	// The HUD window has no session store, so resolve the owning agent's display
-	// name here (for the "opened by X" attribution chip) and stamp it on.
-	let stamped = payload;
-	if (payload.sessionId && !payload.sourceAgent) {
-		const sessions = sessionsStore.get('sessions', []) as Array<{ id?: string; name?: string }>;
-		const sourceAgent = sessions.find((s) => s.id === payload.sessionId)?.name;
-		if (sourceAgent) stamped = { ...payload, sourceAgent };
-	}
-	return deliverCadenzaToHud(mainWindow, cadenzaHudDeps, stamped);
-}
-
-/** Host views are a bridge between two opt-in Encore features. Re-read both
- * flags on every mutation: a disabled feature must never retain a pending view. */
-function arePluginHostViewsEnabled(): boolean {
-	const features = store.get('encoreFeatures', {}) as Record<string, boolean>;
-	return features.plugins === true && features.concerto === true;
-}
-
-/** Forward one host-owned mutation over the exact Concerto renderer channels
- * used by the CLI bridge. No renderer handles or plugin code cross this seam. */
-function forwardPluginHostView(mutation: HostViewMutation): boolean {
-	if (!(mutation.kind === 'remove' && mutation.force) && !arePluginHostViewsEnabled()) return false;
-	if (!mainWindow || mainWindow.isDestroyed() || !isWebContentsAvailable(mainWindow)) return false;
-	const targetWindow = mainWindow;
-	const sourcePlugin =
-		pluginManager?.getRegistry().records.find((record) => record.id === mutation.view.pluginId)
-			?.manifest?.name ?? mutation.view.pluginId;
-	return forwardPluginHostViewToRenderer(mutation, {
-		sourcePlugin,
-		isCadenzaEnabled: store.get('encoreFeatures', {})?.concerto === true,
-		sendToMain: (channel, payload) => targetWindow.webContents.send(channel, payload),
-		deliverCadenza,
-		deliverCadenzaToExistingHud,
-	});
-}
-
-const pluginHostViews = new PluginHostViewRegistry({
-	isEnabled: arePluginHostViewsEnabled,
-	getHostViews: () => pluginManager?.getContributions().hostViews ?? [],
-	isPluginRecordPresent: (pluginId) =>
-		pluginManager?.getRegistry().records.some((record) => record.id === pluginId) ?? false,
-	forward: forwardPluginHostView,
+// See src/main/cadenza-bridge/ and src/main/plugin-host-view-bridge/ for what
+// each of these does (Phase 5 refactoring).
+const { deliverCadenza } = createCadenzaDelivery({
+	getMainWindow: () => mainWindow,
+	sessionsStore,
+	settingsStore: store,
+	cadenzaHudDeps,
 });
+const { arePluginHostViewsEnabled, pluginHostViews } = createPluginHostViewBridge({
+	getMainWindow: () => mainWindow,
+	getPluginManager: () => pluginManager,
+	settingsStore: store,
+	deliverCadenza,
+});
+registerCadenzaIpcHandlers({ getMainWindow: () => mainWindow, settingsStore: store });
 
 // Disabling either side of the bridge purges any live views immediately. If both
 // are enabled after a flag change, re-sync static data without asking a renderer
@@ -746,37 +527,6 @@ store.onDidChange('encoreFeatures', (encoreFeatures) => {
 		return;
 	}
 	pluginHostViews.sync();
-});
-
-// A `decision` cadenza's chosen option replies to the owning agent: inject the
-// value as a live prompt into that agent's session via the main renderer's
-// existing remote-command path (the same one `maestro-cli dispatch` uses). The
-// agent process is already spawned (with SSH if configured), so feeding its live
-// session inherits that transport - no new spawn, no separate SSH handling.
-ipcMain.on('cadenza-hud:decision', (_event, sessionId: string, message: string) => {
-	// Same Concerto gate as the other cadenza entry points: with the flag off no
-	// decision card can exist, so a decision arriving anyway must not inject a
-	// prompt into a live agent session.
-	if (store.get('encoreFeatures')?.concerto !== true) return;
-	if (!mainWindow || mainWindow.isDestroyed()) return;
-	if (!sessionId || !message) return;
-	// force=true (5th arg): a decision card is answered mid-turn, so the owning
-	// agent is busy by definition; without the force flag the renderer's busy
-	// guard would silently drop the choice while the UI reports it was sent.
-	mainWindow.webContents.send('remote:executeCommand', sessionId, message, 'ai', undefined, true);
-});
-
-// A chat "point" chip that targets a cadenza asks main to pulse it. Cadenzas live
-// in the HUD renderer (a separate window with its own store), so the flash must be
-// routed to whichever renderer actually holds the card: the HUD window when it's
-// up, otherwise the main window (the in-app fallback layer). Gated by Concerto so
-// it's inert when off (no cadenzas exist then anyway).
-ipcMain.on('cadenza:flash', (_event, id: string) => {
-	if (!id) return;
-	if (store.get('encoreFeatures')?.concerto !== true) return;
-	const hud = getCadenzaHudWindow();
-	const target = hud && !hud.isDestroyed() ? hud : mainWindow;
-	if (target && !target.isDestroyed()) target.webContents.send('remote:cadenzaFlash', id);
 });
 
 // Create web server factory with dependency injection (Phase 2 refactoring)
@@ -1575,50 +1325,15 @@ app
 			},
 		});
 
-		// Initialize the Pianola supervised daemon. It owns Pianola's background
-		// watchers and orchestrations as supervised child processes (restart on
-		// crash, relaunch on app start, visible health), replacing the unmanaged
-		// nohup model. It self-gates on encoreFeatures.pianola and reconciles from a
-		// shared store file that both the CLI and renderer write.
-		pianolaSupervisor = new PianolaSupervisor({
-			isEnabled: () => {
-				const ef = store.get('encoreFeatures', {}) as Record<string, boolean>;
-				return ef.pianola === true;
-			},
-			getPianolaAgentId: () => {
-				const sessions = sessionsStore.get('sessions', []) as Array<{
-					id?: string;
-					isPianola?: boolean;
-				}>;
-				return sessions.find((s) => s?.isPianola === true)?.id;
-			},
+		// Initialize the Pianola supervised daemon and its scheduled re-learn job.
+		// See src/main/pianola/pianola-lifecycle.ts for what each one does.
+		const pianolaLifecycle = createPianolaLifecycle({
+			settingsStore: store,
+			sessionsStore,
+			logger,
 		});
-
-		// Pianola scheduled re-learn: keeps the learned profile fresh as a PROPOSAL
-		// (stages suggestions; never overwrites the live profile/rules) and
-		// relaunches stale supervised targets, on a fixed cadence. Self-gates per
-		// tick on encoreFeatures.pianola. Mining reuses the existing `pianola learn`
-		// crawler via the bundled CLI; the composition is pure with injected deps.
-		pianolaRelearnScheduler = new PianolaRelearnScheduler({
-			isEnabled: () => {
-				const ef = store.get('encoreFeatures', {}) as Record<string, boolean>;
-				return ef.pianola === true;
-			},
-			runJob: async () => {
-				await runRelearnJob({
-					isEnabled: () => {
-						const ef = store.get('encoreFeatures', {}) as Record<string, boolean>;
-						return ef.pianola === true;
-					},
-					mine: mineDecisionPairsViaCli,
-					readExisting: readExistingForRelearn,
-					writeSuggestions,
-					relaunchStale: () => pianolaSupervisor?.relaunchStale() ?? 0,
-					now: Date.now,
-					log: (line) => logger.info(line, '[PianolaRelearn]'),
-				});
-			},
-		});
+		pianolaSupervisor = pianolaLifecycle.supervisor;
+		pianolaRelearnScheduler = pianolaLifecycle.relearnScheduler;
 
 		// Plugin manager: discovers installed community plugins, tracks their
 		// enable state, verifies signatures, and (tier 1) runs their sandboxed
@@ -2935,11 +2650,56 @@ app
 
 		// Set up IPC handlers
 		logger.debug('Setting up IPC handlers', 'Startup');
-		setupIpcHandlers();
+		setupIpcHandlers({
+			getMainWindow: () => mainWindow,
+			getProcessManager: () => processManager,
+			getWebServer: () => webServer,
+			setWebServer: (server) => {
+				webServer = server;
+			},
+			getAgentDetector: () => agentDetector,
+			getCueEngine: () => cueEngine,
+			getPianolaSupervisor: () => pianolaSupervisor,
+			getPluginManager: () => pluginManager,
+			getPluginSandboxHost: () => pluginSandboxHost,
+			getPluginGroupingRegistry: () => pluginGroupingRegistry,
+			getPluginAuthStore: () => pluginAuthStore,
+			getPluginEventBus: () => pluginEventBus,
+			getInteractiveReplayController: () => interactiveReplayController,
+			setNoteSessionActivated: (fn) => {
+				noteSessionActivatedInPersistence = fn;
+			},
+			app,
+			settingsStore: store,
+			sessionsStore,
+			groupsStore,
+			agentConfigsStore,
+			windowStateStore,
+			claudeSessionOriginsStore,
+			agentSessionOriginsStore,
+			bootstrapStore,
+			safeSend,
+			windowRegistry,
+			windowManager,
+			createWebServer,
+			wakatimeManager,
+			maestroCliManager,
+			getAgentConfigForAgent,
+			getCustomEnvVarsForAgent,
+		});
 
 		// Set up process event listeners
 		logger.debug('Setting up process event listeners', 'Startup');
-		setupProcessListeners();
+		wireProcessListeners({
+			getProcessManager: () => processManager,
+			getWebServer: () => webServer,
+			getAgentDetector: () => agentDetector,
+			getCueEngine: () => cueEngine,
+			getPluginEventBus: () => pluginEventBus,
+			safeSend,
+			settingsStore: store,
+			wakatimeManager,
+		});
 
 		// Wire agent-run lifecycle capture to the ProcessManager (F1). Always-on
 		// per D1: minimal metadata capture is observability, not an opt-in feature.
@@ -3273,531 +3033,3 @@ quitHandler = createQuitHandler({
 	getWindowRegistry: () => windowRegistry,
 });
 quitHandler.setup();
-
-// startCliActivityWatcher is now handled by cliWatcher (Phase 4 refactoring)
-
-function setupIpcHandlers() {
-	// Settings, sessions, and groups persistence - extracted to src/main/ipc/handlers/persistence.ts
-
-	// Web/Live handlers - extracted to src/main/ipc/handlers/web.ts
-	registerWebHandlers({
-		getWebServer: () => webServer,
-		setWebServer: (server) => {
-			webServer = server;
-		},
-		createWebServer,
-		settingsStore: store,
-	});
-
-	// Git operations - extracted to src/main/ipc/handlers/git.ts
-	registerGitHandlers({
-		settingsStore: store,
-		getMainWindow: () => mainWindow,
-	});
-
-	// Auto Run operations - extracted to src/main/ipc/handlers/autorun.ts
-	registerAutorunHandlers({
-		mainWindow,
-		getMainWindow: () => mainWindow,
-		app,
-		settingsStore: store,
-	});
-
-	// Playbook operations - extracted to src/main/ipc/handlers/playbooks.ts
-	registerPlaybooksHandlers({
-		mainWindow,
-		getMainWindow: () => mainWindow,
-		app,
-	});
-
-	// History operations - extracted to src/main/ipc/handlers/history.ts
-	// Uses HistoryManager singleton for per-session storage
-	registerHistoryHandlers({
-		safeSend,
-		emitPluginEvent: (event) => pluginEventBus?.emit(event),
-		getMaxEntries: () => store.get('maxLogBuffer', 5000) as number,
-		getSshRemoteById,
-		getSessionById: (id: string) => {
-			const sessions = (sessionsStore.get('sessions', []) as Array<Record<string, unknown>>).filter(
-				(s) => typeof s === 'object' && s !== null
-			);
-			return sessions.find((s) => s.id === id);
-		},
-	});
-
-	// Director's Notes - unified history + synopsis generation
-	registerDirectorNotesHandlers({
-		getProcessManager: () => processManager,
-		getAgentDetector: () => agentDetector,
-		agentConfigsStore,
-		getMainWindow: () => mainWindow,
-	});
-
-	// Cross-agent @mention dispatch - streams a target agent's response back
-	// into the source agent's transcript (Phase 03).
-	registerCrossAgentHandlers({
-		getProcessManager: () => processManager,
-		getAgentDetector: () => agentDetector,
-		sessionsStore,
-		agentConfigsStore,
-		settingsStore: store,
-		sshStore: createSshRemoteStoreAdapter(store),
-		getCustomEnvVars: getCustomEnvVarsForAgent,
-		safeSend,
-	});
-
-	// Cue - event-driven automation engine
-	registerCueHandlers({
-		getCueEngine: () => cueEngine,
-	});
-
-	// Cue Backup - snapshot / restore .maestro/cue.yaml + prompts (Cue modal Backup tab)
-	registerCueBackupHandlers({
-		sessionsStore,
-	});
-
-	// Agent management operations - extracted to src/main/ipc/handlers/agents.ts
-	registerAgentsHandlers({
-		getAgentDetector: () => agentDetector,
-		agentConfigsStore,
-		settingsStore: store,
-		sessionsStore,
-	});
-
-	// Process management operations - extracted to src/main/ipc/handlers/process.ts
-	registerProcessHandlers({
-		getProcessManager: () => processManager,
-		getAgentDetector: () => agentDetector,
-		agentConfigsStore,
-		settingsStore: store,
-		getMainWindow: () => mainWindow,
-		safeSend,
-		sessionsStore,
-		interactiveReplayController: interactiveReplayController ?? undefined,
-		getCueProcesses: () => {
-			// Always query the executor's active process map - processes may still be
-			// running even if the engine has been disabled (in-flight runs complete
-			// independently of engine state).
-			const processList = getCueProcessList();
-			if (processList.length === 0) return [];
-			const activeRuns = cueEngine?.getActiveRuns() ?? [];
-			// Merge PID/command data from executor with metadata from run manager
-			return processList.map((proc) => {
-				const run = activeRuns.find((r) => r.runId === proc.runId);
-				return {
-					...proc,
-					sessionName: run?.sessionName ?? '',
-					subscriptionName: run?.subscriptionName ?? '',
-					eventType: run?.event.type ?? '',
-				};
-			});
-		},
-	});
-
-	// Persistence operations - extracted to src/main/ipc/handlers/persistence.ts
-	const persistenceHandlers = registerPersistenceHandlers({
-		settingsStore: store,
-		sessionsStore,
-		groupsStore,
-		getWebServer: () => webServer,
-		// Metadata-only session/agent lifecycle -> subscribed plugins. Null-safe:
-		// the bus is created during plugin init and re-authorizes every delivery
-		// against live grants, so this is a no-op when plugins are disabled.
-		emitPluginEvent: (event) => pluginEventBus?.emit(event),
-	});
-	// Wire the plugin focus verbs into the persistence layer's session.activated
-	// dedupe so the two emit paths share one last-emitted id.
-	noteSessionActivatedInPersistence = persistenceHandlers.noteSessionActivated;
-
-	// System operations - extracted to src/main/ipc/handlers/system.ts
-	registerSystemHandlers({
-		getMainWindow: () => mainWindow,
-		app,
-		settingsStore: store,
-		tunnelManager,
-		getWebServer: () => webServer,
-		bootstrapStore, // For iCloud/sync settings
-	});
-
-	// Claude Code sessions - extracted to src/main/ipc/handlers/claude.ts
-	registerClaudeHandlers({
-		claudeSessionOriginsStore,
-		getMainWindow: () => mainWindow,
-	});
-
-	// Initialize output parsers for all agents (Codex, OpenCode, Claude Code)
-	// This must be called before any agent output is processed
-	initializeOutputParsers();
-
-	// Initialize session storages and register generic agent sessions handlers
-	// This provides the new window.maestro.agentSessions.* API
-	// Pass the shared claudeSessionOriginsStore so session names/stars are consistent
-	initializeSessionStorages({ claudeSessionOriginsStore });
-	registerAgentSessionsHandlers({ getMainWindow: () => mainWindow, agentSessionOriginsStore });
-
-	// Register Group Chat handlers
-	registerGroupChatHandlers({
-		getMainWindow: () => mainWindow,
-		getProcessManager: () => processManager,
-		getAgentDetector: () => agentDetector,
-		getCustomEnvVars: getCustomEnvVarsForAgent,
-		getAgentConfig: getAgentConfigForAgent,
-	});
-
-	// Register Debug Package handlers
-	registerDebugHandlers({
-		getMainWindow: () => mainWindow,
-		getAgentDetector: () => agentDetector,
-		getProcessManager: () => processManager,
-		getWebServer: () => webServer,
-		settingsStore: store,
-		sessionsStore,
-		groupsStore,
-		bootstrapStore,
-	});
-
-	// Register Spec Kit handlers (no dependencies needed)
-	registerSpeckitHandlers();
-
-	// Register OpenSpec handlers (no dependencies needed)
-	registerOpenSpecHandlers();
-
-	// Register BMAD handlers (no dependencies needed)
-	registerBmadHandlers();
-
-	// Register Core Prompts handlers (no dependencies needed)
-	registerPromptsHandlers();
-
-	// Register project Memory handlers (Claude Code per-project memory viewer)
-	registerMemoryHandlers();
-
-	// Register Pianola handlers (autonomous manager: rules, decisions, and the
-	// supervised daemon). The supervisor is constructed during core-service init
-	// above, so it is available here; guard anyway to keep types honest.
-	if (pianolaSupervisor) {
-		registerPianolaHandlers({
-			settingsStore: store,
-			supervisor: pianolaSupervisor,
-		});
-	}
-
-	// Register Plugins handlers (community plugin subsystem, list-only in Phase 0).
-	// The manager is constructed during core-service init above; guard for types.
-	if (pluginManager && pluginAuthStore) {
-		registerPluginsHandlers({
-			settingsStore: store,
-			manager: pluginManager,
-			sandboxHost: pluginSandboxHost ?? undefined,
-			authStore: pluginAuthStore,
-			groupingRegistry: pluginGroupingRegistry ?? undefined,
-		});
-	}
-
-	// Register AgentRun control-plane handlers (neutral run/campaign ledger).
-	registerAgentRunHandlers({
-		getProcessManager: () => processManager,
-		settingsStore: store,
-	});
-
-	// Register Browser Session handlers (clear per-partition browsing data)
-	registerBrowserSessionHandlers();
-
-	// Register Coworking handlers + start the IPC bridge socket and refresh the bundled
-	// MCP-server script. The bridge runs whenever Maestro is up; per-agent activation
-	// is opt-in via Settings → Encore Features → Coworking Setup. Bridge startup is
-	// non-fatal - feature degrades to "not available" until next launch.
-	registerCoworkingHandlers({ getMainWindow: () => mainWindow });
-	void (async () => {
-		try {
-			await ensureCoworkingServerScript();
-			await startCoworkingBridge({
-				resolveSessionFromPid: (pid) =>
-					resolveSessionFromPidWalk(
-						pid,
-						(candidate) => processManager?.getSessionIdByPid(candidate) ?? null
-					),
-			});
-		} catch (err) {
-			// EADDRINUSE means another Maestro process already owns the bridge
-			// socket/pipe for this userData slug. Bridge startup is deliberately
-			// non-fatal (the feature degrades to "not available" until the next
-			// launch), so a second instance losing the race is an expected
-			// outcome rather than a defect worth reporting (MAESTRO-WH).
-			const code = (err as NodeJS.ErrnoException | null)?.code;
-			if (code !== 'EADDRINUSE') {
-				void captureException(err instanceof Error ? err : new Error(String(err)), {
-					operation: 'startup:coworkingBridge',
-				});
-			}
-			logger.warn(`Failed to start coworking bridge: ${String(err)}`, 'Startup');
-		}
-	})();
-	// Register multi-window handlers (windows:* channel surface). Registered here
-	// because the running app wires handlers through setupIpcHandlers(), not
-	// registerAllHandlers(). The registry and window manager are module-scope
-	// instances; lazy getters resolve the live instance at call time.
-	registerWindowsHandlers({
-		getWindowRegistry: () => windowRegistry,
-		getWindowManager: () => windowManager,
-	});
-	// Push registry ownership moves out to every window so each renderer's
-	// WindowContext can refresh which agents it surfaces (and the Left Bar's
-	// cross-window badges). The registry is a module-scope instance, so pass it
-	// directly rather than through the handlers' lazy getter.
-	wireWindowRegistryBroadcast(windowRegistry);
-	// Close a secondary window as soon as its last agent moves out - an empty
-	// secondary shell can surface nothing (every agent is owned by some window),
-	// so the agent-level move flow tidies it up automatically.
-	wireEmptySecondaryWindowAutoClose(windowRegistry);
-	// Persist a window rename or a panel-collapse toggle as soon as it happens
-	// (rather than only on quit), so both survive even an abrupt exit. A panel
-	// toggle fires no window move/resize, so without this its saved value would go
-	// stale. saveWindowState snapshots the whole live registry, so passing the
-	// affected window's id is enough.
-	windowRegistry.onChange((change) => {
-		if ((change.type === 'name-changed' || change.type === 'panel-changed') && change.windowId) {
-			saveWindowState(windowStateStore, windowRegistry, change.windowId);
-		}
-	});
-
-	// Record aggregate multi-window usage telemetry (secondary windows opened +
-	// peak concurrent windows) as windows open. Gated on the user's
-	// `statsCollectionEnabled` analytics setting; records nothing when off, and a
-	// stats failure can never break window creation (see wireMultiWindowTelemetry).
-	wireMultiWindowTelemetry(windowRegistry, { settingsStore: store });
-	// Register Context Merge handlers for session context transfer and grooming
-	registerContextHandlers({
-		getMainWindow: () => mainWindow,
-		getProcessManager: () => processManager,
-		getAgentDetector: () => agentDetector,
-		agentConfigsStore,
-	});
-
-	// Register Marketplace handlers for fetching and importing playbooks
-	registerMarketplaceHandlers({
-		app,
-		settingsStore: store,
-		getMainWindow: () => mainWindow,
-	});
-
-	// Register Stats handlers for usage tracking
-	registerStatsHandlers({
-		getMainWindow: () => mainWindow,
-		settingsStore: store,
-	});
-
-	// Register Cue Stats handlers for the Cue Dashboard aggregation query.
-	// Pass `getCueEngine` so the handler can fall back to the live cue config
-	// when persisted `pipeline_id` is null (legacy events / events recorded
-	// before lineage tracking was enabled).
-	registerCueStatsHandlers({
-		settingsStore: store,
-		getCueEngine: () => cueEngine,
-	});
-
-	// Register Document Graph handlers for file watching
-	registerDocumentGraphHandlers({
-		getMainWindow: () => mainWindow,
-		app,
-	});
-
-	// Register SSH Remote handlers for managing SSH configurations
-	registerSshRemoteHandlers({
-		settingsStore: store,
-	});
-
-	// Set up callback for group chat router to lookup sessions for auto-add @mentions
-	setGetSessionsCallback(() => {
-		const sessions = sessionsStore.get('sessions', []);
-		return sessions.map((s: any) => {
-			// Resolve SSH remote name if session has SSH config
-			let sshRemoteName: string | undefined;
-			if (s.sessionSshRemoteConfig?.enabled && s.sessionSshRemoteConfig.remoteId) {
-				const sshConfig = getSshRemoteById(s.sessionSshRemoteConfig.remoteId);
-				sshRemoteName = sshConfig?.name;
-			}
-			return {
-				id: s.id,
-				name: s.name,
-				toolType: s.toolType,
-				cwd: s.cwd || s.fullPath || os.homedir(),
-				customArgs: s.customArgs,
-				customEnvVars: s.customEnvVars,
-				customModel: s.customModel,
-				// Claude token-source selection, so group chat participants honor
-				// the same maestro-p TUI / API / dynamic choice as their agent.
-				enableMaestroP: s.enableMaestroP,
-				maestroPMode: s.maestroPMode,
-				maestroPPath: s.maestroPPath,
-				sshRemoteName,
-				// Pass full SSH config for remote execution support
-				sshRemoteConfig: s.sessionSshRemoteConfig,
-				autoRunFolderPath: s.autoRunFolderPath,
-				worktreeBasePath: s.worktreeConfig?.basePath,
-			};
-		});
-	});
-
-	// Set up callback for group chat router to lookup custom env vars for agents
-	setGetCustomEnvVarsCallback(getCustomEnvVarsForAgent);
-	setGetAgentConfigCallback(getAgentConfigForAgent);
-
-	// Set up callback for group chat router to get moderator conductor profile
-	setGetModeratorSettingsCallback(() => ({
-		conductorProfile: (store.get('conductorProfile', '') as string) || '',
-	}));
-
-	// Set up SSH store for group chat SSH remote execution support
-	setSshStore(createSshRemoteStoreAdapter(store));
-
-	// Set up callback for group chat to get custom shell path (for Windows PowerShell preference)
-	// This is used by both group-chat-router.ts and group-chat-agent.ts via the shared config module
-	const getCustomShellPathFn = () => store.get('customShellPath', '') as string | undefined;
-	setGetCustomShellPathCallback(getCustomShellPathFn);
-
-	// Setup logger event forwarding to renderer
-	setupLoggerEventForwarding(() => mainWindow);
-
-	// Register filesystem handlers (extracted to handlers/filesystem.ts)
-	registerFilesystemHandlers();
-
-	// System operations (dialog, fonts, shells, tunnel, devtools, updates, logger)
-	// extracted to src/main/ipc/handlers/system.ts
-
-	// Claude Code sessions - extracted to src/main/ipc/handlers/claude.ts
-
-	// Agent Error Handling API - extracted to src/main/ipc/handlers/agent-error.ts
-	registerAgentErrorHandlers();
-
-	// Register notification handlers (extracted to handlers/notifications.ts)
-	registerNotificationsHandlers({ getMainWindow: () => mainWindow });
-
-	// Register attachments handlers (extracted to handlers/attachments.ts)
-	registerAttachmentsHandlers({ app });
-
-	// Register leaderboard handlers (extracted to handlers/leaderboard.ts)
-	registerLeaderboardHandlers({
-		app,
-		settingsStore: store,
-	});
-
-	// Register Symphony handlers for token donation / open source contributions
-	registerSymphonyHandlers({
-		app,
-		getMainWindow: () => mainWindow,
-		sessionsStore,
-		settingsStore: store,
-	});
-
-	// Register tab naming handlers for automatic tab naming
-	registerTabNamingHandlers({
-		getProcessManager: () => processManager,
-		getAgentDetector: () => agentDetector,
-		agentConfigsStore,
-		settingsStore: store,
-	});
-
-	// Register WakaTime handlers (CLI check, API key validation)
-	registerWakatimeHandlers(wakatimeManager);
-
-	// Register Maestro CLI handlers (status check + install/update)
-	registerMaestroCliHandlers(maestroCliManager);
-
-	// Register feedback handlers (gh auth + feedback submission)
-	registerFeedbackHandlers({
-		getProcessManager: () => processManager,
-		debugPackageDeps: {
-			getAgentDetector: () => agentDetector,
-			getProcessManager: () => processManager,
-			getWebServer: () => webServer,
-			settingsStore: store,
-			sessionsStore,
-			groupsStore,
-			bootstrapStore,
-		},
-	});
-}
-
-// Handle process output streaming (set up after initialization)
-// Phase 3 refactoring - delegates to extracted process-listeners module
-function setupProcessListeners() {
-	if (processManager) {
-		setupProcessListenersModule(processManager, {
-			getProcessManager: () => processManager,
-			getWebServer: () => webServer,
-			getAgentDetector: () => agentDetector,
-			safeSend,
-			powerManager,
-			groupChatEmitters,
-			emitPluginEvent: (event) => pluginEventBus?.emit(event),
-			groupChatRouter: {
-				routeModeratorResponse,
-				routeAgentResponse,
-				markParticipantResponded,
-				spawnModeratorSynthesis,
-				getGroupChatReadOnlyState,
-				respawnParticipantWithRecovery,
-				clearActiveParticipantTaskSession,
-				clearModeratorResponseTimeout,
-			},
-			groupChatStorage: {
-				loadGroupChat,
-				updateGroupChat,
-				updateParticipant,
-			},
-			sessionRecovery: {
-				needsSessionRecovery,
-				initiateSessionRecovery,
-			},
-			outputBuffer: {
-				appendToGroupChatBuffer,
-				getGroupChatBufferedOutput,
-				clearGroupChatBuffer,
-			},
-			outputParser: {
-				extractTextFromStreamJson,
-				parseParticipantSessionId,
-			},
-			usageAggregator: {
-				calculateContextTokens,
-			},
-			getStatsDB,
-			debugLog,
-			patterns: {
-				REGEX_MODERATOR_SESSION,
-				REGEX_MODERATOR_SESSION_TIMESTAMP,
-				REGEX_AI_SUFFIX,
-				REGEX_AI_TAB_ID,
-				REGEX_BATCH_SESSION,
-				REGEX_SYNOPSIS_SESSION,
-			},
-			logger,
-			getCueEngine: () => cueEngine,
-			isCueEnabled: () => {
-				const ef = store.get('encoreFeatures', {}) as Record<string, boolean>;
-				return !!ef.maestroCue;
-			},
-			getSshRemoteByName: (name: string) => {
-				const remotes = store.get('sshRemotes', []);
-				return remotes.find((r) => r.name === name) ?? null;
-			},
-			getAgentContextWindow: (agentId: string) => {
-				// Prefer a runtime-discovered context window from the capability
-				// snapshot if one was probed. Falls back to the static table and
-				// finally to the agent definition's configOption default.
-				const snapshot = capabilitySnapshots.get(agentId);
-				if (typeof snapshot?.contextWindow === 'number' && snapshot.contextWindow > 0) {
-					return snapshot.contextWindow;
-				}
-				const def = getAgentDefinition(agentId);
-				const contextOpt = def?.configOptions?.find((o) => o.key === 'contextWindow');
-				const fallbackDefault =
-					typeof contextOpt?.default === 'number' ? contextOpt.default : FALLBACK_CONTEXT_WINDOW;
-				return DEFAULT_CONTEXT_WINDOWS[agentId as AgentId] ?? fallbackDefault;
-			},
-		});
-
-		// WakaTime heartbeat listener (query-complete → heartbeat, exit → cleanup)
-		setupWakaTimeListener(processManager, wakatimeManager, store);
-	}
-}
