+++ src/main/plugin-host-view-bridge/index.ts
@@ -1,0 +1,2 @@
+export { createPluginHostViewBridge } from './bridge';
+export type { PluginHostViewBridgeDependencies } from './types';
